+++ tests/test_insert_timestamp.py
@@ -53,82 +53,6 @@ class ViewsTest(StaticLiveServerTestCase):
     def tearDown(self):
         self.driver.quit()
 
-    def test_register(self):
-        self.driver.get(urljoin(self.live_server_url, 'register'))
-        self.driver.find_element_by_id('input_txt_username').send_keys('new_user')
-        self.driver.find_element_by_id('input_txt_email').send_keys('new_user@example.com')
-        self.driver.find_element_by_id('input_txt_password').send_keys('new_password_1234')
-        self.driver.find_element_by_id('input_txt_confirmation').send_keys('new_password_1234')
-        self.driver.find_element_by_id('input_btn_register').click()
-        try:
-            user = User.objects.get(username='new_user')
-            self.assertEqual(user.username, 'new_user')
-            self.assertEqual(user.email, 'new_user@example.com')
-            self.assertEqual(len(user.password), 88)
-            self.assertTrue(not user.is_superuser)
-            self.assertTrue(not user.is_staff)
-            self.assertTrue(user.is_active)
-            self.assertEqual(self.driver.current_url.rstrip('/'), self.live_server_url)
-        except User.DoesNotExist:
-            self.assertTrue(False)
-
-    def test_login(self):
-        self.login(driver=self.driver,
-                   live_server_url=self.live_server_url,
-                   username='default_user',
-                   password='default_password_1234')
-
-        self.assertEqual(self.driver.current_url.rstrip('/'), self.live_server_url)
-
-    def test_log_out(self):
-        self.login(driver=self.driver,
-                   live_server_url=self.live_server_url,
-                   username='default_user',
-                   password='default_password_1234')
-
-        link = self.driver.find_element_by_id('a_logout')
-        link.click()
-        self.assertEqual(self.driver.current_url.rstrip('/'), self.live_server_url)
-
-    def test_combine_tracks(self):
-        self.login(driver=self.driver,
-                   live_server_url=self.live_server_url,
-                   username='default_user',
-                   password='default_password_1234')
-
-        # Remove previous testing files
-        for file in glob(os.path.join(self.downloads_dir, 'TrackEditor_combine_tracks*.gpx')):
-            os.remove(file)
-
-        self.driver.get(urljoin(self.live_server_url, 'combine_tracks'))
-
-        self.driver.\
-            find_element_by_id('select-file-1').\
-            send_keys(os.path.join(self.test_path,
-                                   'samples',
-                                   'island_1.gpx'))
-
-        self.driver.\
-            find_element_by_id('select-file-2').\
-            send_keys(os.path.join(self.test_path,
-                                   'samples',
-                                   'island_2.gpx'))
-
-        self.driver.find_element_by_id('input_btn_combine').click()
-        self.driver.find_element_by_id('input_btn_download').click()
-        time.sleep(2)  # wait download
-        downloaded_file = \
-            glob(os.path.join(self.downloads_dir, 'TrackEditor_combine_tracks*.gpx'))[-1]
-
-        self.assertEqual(
-            md5sum(downloaded_file),
-            md5sum(os.path.join(self.test_path,
-                                'references',
-                                'test_combine_tracks.gpx')
-                   )
-        )
-        self.assertIsNotNone(self.driver.find_element_by_id('js-map'))
-
     def test_insert_time(self):
         self.login(driver=self.driver,
                    live_server_url=self.live_server_url,
@@ -175,118 +99,6 @@ class ViewsTest(StaticLiveServerTestCase):
                    )
         )
         self.assertIsNotNone(self.driver.find_element_by_id('js-map'))
-
-    def test_editor_no_logged(self):
-        self.driver.get(self.live_server_url)
-
-        link = self.driver.find_element_by_id('a_editor')
-        link.click()
-
-        warning_msg = self.driver.find_element_by_id('div_warning_msg')
-
-        self.assertIn('only available for users', warning_msg.text)
-
-        self.assertEqual(self.driver.current_url.rstrip('/'),
-                         urljoin(self.live_server_url, 'users_only'))
-
-    def test_editor_logged(self):
-        self.login(driver=self.driver,
-                   live_server_url=self.live_server_url,
-                   username='default_user',
-                   password='default_password_1234')
-
-        self.driver.get(self.live_server_url)
-
-        link = self.driver.find_element_by_id('a_editor')
-        link.click()
-        self.assertEqual(self.driver.current_url.rstrip('/'),
-                         urljoin(self.live_server_url, 'editor'))
-
-
-class CombineTracksTest(StaticLiveServerTestCase):
-    def setUp(self):
-        options = webdriver.ChromeOptions()
-        options.headless = True
-        self.driver = webdriver.Chrome(chrome_options=options)
-        self.test_path = os.path.dirname(__file__)
-        self.driver.get(urljoin(self.live_server_url, 'combine_tracks'))
-
-    def tearDown(self):
-        self.driver.quit()
-
-    def test_upload_too_many_tracks(self):
-        for i in range(1, 6):
-            self.driver.\
-                find_element_by_id(f'select-file-{i}').\
-                send_keys(os.path.join(self.test_path,
-                                       'samples',
-                                       f'island_{i}.gpx'))
-        self.driver.\
-            find_element_by_id('select-file-6').\
-            send_keys(os.path.join(self.test_path,
-                                   'samples',
-                                   'basic_sample.gpx'))
-
-        error_msg = self.driver.find_element_by_id('div_error_msg_js')
-
-        self.assertEqual(error_msg.text, 'No more than 5 files are allowed.')
-
-    def test_upload_wrong_extension(self):
-        self.driver.\
-            find_element_by_id('select-file-1').\
-            send_keys(os.path.join(self.test_path,
-                                   'samples',
-                                   'wrong_extension.txt'))
-
-        error_msg = self.driver.find_element_by_id('div_error_msg_js')
-
-        self.assertEqual(
-            error_msg.text,
-            "Extension for wrong_extension.txt is not valid ['gpx']")
-
-    def test_upload_repeated_files(self):
-        for i in range(1, 3):
-            self.driver. \
-                find_element_by_id(f'select-file-{i}'). \
-                send_keys(os.path.join(self.test_path,
-                                       'samples',
-                                       'island_1.gpx'))
-
-        error_msg = self.driver.find_element_by_id('div_error_msg_js')
-
-        self.assertEqual(
-            error_msg.text,
-            'Repeated file is selected: island_1.gpx')
-
-    def test_upload_big_file(self):
-        self.driver. \
-            find_element_by_id('select-file-1'). \
-            send_keys(os.path.join(self.test_path,
-                                   'samples',
-                                   'over_10mb.gpx'))
-
-        error_msg = self.driver.find_element_by_id('div_error_msg_js')
-
-        self.assertEqual(
-            error_msg.text,
-            'File over_10mb.gpx is 14 Mb. It must be smaller than 10 Mb')
-
-    def test_combine_no_file(self):
-        self.driver.find_element_by_id('input_btn_combine').click()
-        warning_msg = self.driver.find_element_by_id('div_warning_msg')
-
-        self.assertEqual(warning_msg.text, 'No file has been selected.')
-
-    def test_bad_formed_file(self):
-        self.driver. \
-            find_element_by_id('select-file-1'). \
-            send_keys(os.path.join(self.test_path,
-                                   'samples',
-                                   'bad_formed.gpx'))
-
-        self.driver.find_element_by_id('input_btn_combine').click()
-        error_msg = self.driver.find_element_by_id('div_error_msg')
-        self.assertEqual(error_msg.text, 'Error loading files')
 
 
 class InsertTimestampTest(StaticLiveServerTestCase):
